requirements.txt
Django>=5.0.0
gunicorn
whitenoise
dj-database-url
psycopg2-binary
Pillow
numpy
pandas
scikit-learn
requests
mysqlclient
PyYAML
python-dotenv
